+++ internal/app/admin/http/cache.go
@@ -1,6 +1,7 @@
 package handler
 
 import (
+	"fmt"
 	"net/http"
 
 	v2 "github.com/envoyproxy/go-control-plane/envoy/api/v2"
@@ -78,7 +79,52 @@ func edsDumpHandler(o *orchestrator.Orchestrator) http.HandlerFunc {
 			return
 		}
 
+		_, e = w.Write([]byte(x))
+		if e != nil {
+			w.WriteHeader(http.StatusInternalServerError)
+			return
+		}
 		w.WriteHeader(http.StatusOK)
-		_, _ = w.Write([]byte(x))
+	}
+}
+
+func keyDumpHandler(o *orchestrator.Orchestrator) http.HandlerFunc {
+	return func(w http.ResponseWriter, req *http.Request) {
+		allKeys, err := orchestrator.Orchestrator.GetDownstreamAggregatedKeys(*o)
+		if err != nil {
+			w.WriteHeader(http.StatusInternalServerError)
+			errMessage, _ := stringify.InterfaceToString(&marshallable.Error{
+				Message: fmt.Sprintf("error in getting cache keys: %s", err.Error()),
+			})
+			_, _ = w.Write([]byte(errMessage))
+			return
+		}
+
+		keys := make([]string, 0)
+		for k := range allKeys {
+			keys = append(keys, k)
+		}
+
+		response := &marshallable.Key{
+			Names: keys,
+		}
+		marshalledKeys, err := stringify.InterfaceToString(response)
+		if err != nil {
+			w.WriteHeader(http.StatusInternalServerError)
+			errMessage, _ := stringify.InterfaceToString(&marshallable.Error{
+				Message: fmt.Sprintf("error in marshalling keys: %s", err.Error()),
+			})
+			_, _ = w.Write([]byte(errMessage))
+			return
+		}
+
+		_, e := w.Write([]byte(marshalledKeys))
+		if e != nil {
+			w.WriteHeader(http.StatusInternalServerError)
+			return
+		}
+
+		w.WriteHeader(http.StatusOK)
+
 	}
 }
